# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QilbeeDB Platform API",
    "version": "0.4.0",
    "description": "Durable identity, scoped memory with model-bound semantic vector retrieval, evidence-driven procedural learning and learned-tool development receipts for AI agents."
  },
  "servers": [
    {
      "url": "http://localhost:7474",
      "description": "Local Docker or development server"
    }
  ],
  "security": [
    {
      "platformCredential": []
    }
  ],
...
  "paths": {
...
    "/api/v1/memory/commands": {
      "post": {
        "operationId": "applyMemoryCommand",
        "summary": "Atomically create, update or delete a scoped record",
        "description": "Requires memory_write and an exact grant. Same key and normalized command return the original receipt, including after deletion. Changed commands conflict. No implicit retention quota.",
        "tags": [
          "Memory"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields or unsupported contract version",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or revoked credential",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Capability, tenant or scope is not granted",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No current record in the authorized scope",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency payload or expected revision conflict",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 65,536 bytes",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure or explicit stored-version inconsistency",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCommand"
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/records/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "contract_version",
          "in": "query",
          "required": true,
          "schema": {
            "type": "integer",
            "const": 1
          }
        },
        {
          "name": "project_id",
          "in": "query",
          "required": true,
          "schema": {
            "type": "string",
            "minLength": 1,
            "description": "1\u2013256 UTF-8 bytes, not blank and without control characters."
          }
        },
        {
          "name": "mission_id",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "minLength": 1,
            "description": "1\u2013256 UTF-8 bytes, not blank and without control characters."
          }
        },
        {
          "name": "agent_id",
          "in": "query",
          "required": true,
          "schema": {
            "type": "string",
            "minLength": 1,
            "description": "1\u2013256 UTF-8 bytes, not blank and without control characters."
          }
        },
        {
          "name": "visibility",
          "in": "query",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      ],
      "get": {
        "operationId": "readMemoryRecord",
        "summary": "Read a current unexpired record in the authorized scope",
        "description": "Requires memory_read and an exact grant. Omit mission_id for missionless memory. Tenant and private subject come from the credential.",
        "tags": [
          "Memory"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields or unsupported contract version",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or revoked credential",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Capability, tenant or scope is not granted",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No current record in the authorized scope",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure or explicit stored-version inconsistency",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/query": {
      "post": {
        "operationId": "queryMemory",
        "summary": "Query current scoped records with explicit continuation",
        "description": "Requires memory_read. Case-insensitive substring over primary/secondary/context; exact type/tag filters. UUID ordering, 1,000 results and 10,000 scanned entries per request. Live pagination is not a snapshot or changes feed. No synthetic similarity scores.",
        "tags": [
          "Memory"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields or unsupported contract version",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or revoked credential",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Capability, tenant or scope is not granted",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 65,536 bytes",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure or explicit stored-version inconsistency",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/memory/embeddings": {
      "post": {
        "operationId": "attachMemoryEmbedding",
        "summary": "Attach a model-bound vector to an exact memory revision",
        "description": "Requires memory_write and exact scope. Embedding plus original receipt commits atomically. Source changes invalidate eligibility. Vectors are generated externally; anonymous legacy embeddings are not assigned a model.",
        "tags": [
          "Memory"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields or unsupported contract version",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or revoked credential",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Capability, tenant or scope is not granted",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No current record in the authorized scope",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency payload or expected revision conflict",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 65,536 bytes",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure or explicit stored-version inconsistency",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingCommandRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "platformCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "qdb1 opaque credential",
        "description": "Provision through the trusted local bootstrap command and authenticated credential API."
      }
    },
    "schemas": {
...
      "Visibility": {
        "type": "string",
        "enum": [
          "private",
          "shared"
        ]
      },
...
      "MemoryOperation": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/CreateOperation"
          },
          {
            "$ref": "#/components/schemas/UpdateOperation"
          },
          {
            "$ref": "#/components/schemas/DeleteOperation"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "create": "#/components/schemas/CreateOperation",
            "update": "#/components/schemas/UpdateOperation",
            "delete": "#/components/schemas/DeleteOperation"
          }
        }
      },
      "MemoryCommand": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "idempotency_key": {
            "type": "string",
            "minLength": 1,
            "description": "1\u2013256 UTF-8 bytes; not blank; no control characters. Scoped by authenticated tenant, namespace and subject."
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "operation": {
            "$ref": "#/components/schemas/MemoryOperation"
          }
        },
        "required": [
          "contract_version",
          "idempotency_key",
          "scope",
          "operation"
        ],
        "additionalProperties": false,
        "examples": [
          {
            "contract_version": 1,
            "idempotency_key": "session-42-observation-1",
            "scope": {
              "project_id": "research",
              "mission_id": "mission-42",
              "agent_id": "researcher",
              "visibility": "shared"
            },
            "operation": {
              "type": "create",
              "record": {
                "episode_type": "Observation",
                "event_time_millis": 1700000000000,
                "content": {
                  "primary": "The tool returned three matching documents.",
                  "data": {
                    "document_ids": [
                      "paper-a",
                      "paper-b",
                      "paper-c"
                    ]
                  }
                },
                "tags": [
                  "literature"
                ],
                "metadata": {
                  "source_request_id": "tool-call-17"
                }
              }
            }
          }
        ]
      },
...
      "ReceiptResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "receipt": {
            "$ref": "#/components/schemas/Receipt"
          }
        },
        "required": [
          "contract_version",
          "receipt"
        ],
        "additionalProperties": false
      },
...
      "RecordResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "record": {
            "$ref": "#/components/schemas/MemoryRecord"
          }
        },
        "required": [
          "contract_version",
          "scope",
          "record"
        ],
        "additionalProperties": false
      },
...
      "QueryRequest": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "filter": {
            "$ref": "#/components/schemas/MemoryFilter"
          }
        },
        "required": [
          "contract_version",
          "scope",
          "filter"
        ],
        "additionalProperties": false,
        "examples": [
          {
            "contract_version": 1,
            "scope": {
              "project_id": "research",
              "mission_id": "mission-42",
              "agent_id": "researcher",
              "visibility": "shared"
            },
            "filter": {
              "limit": 20,
              "text_contains": "documents"
            }
          }
        ]
      },
...
      "QueryResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "page": {
            "$ref": "#/components/schemas/QueryPage"
          }
        },
        "required": [
          "contract_version",
          "scope",
          "page"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "method_not_allowed",
                  "invalid_request",
                  "unsupported_contract_version",
                  "revision_conflict",
                  "record_not_found",
                  "idempotency_conflict",
                  "storage_inconsistency",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "contract_version",
          "error"
        ],
        "additionalProperties": false
      },
...
      "EmbeddingResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "receipt": {
            "$ref": "#/components/schemas/EmbeddingReceipt"
          }
        },
        "required": [
          "contract_version",
          "receipt"
        ],
        "additionalProperties": false
      },
...
    }
  }
}